HTML Work/requirements.txt
Flask==1.1.2
google-cloud-storage==1.26.0
gunicorn==20.0.4
pandas==1.0.1
folium==0.10.1
json5==0.9.1
branca==0.4.0